Update Duplicates.java
Found a way to remove the duplicates.  After sorting the list, i check every element and the element next to it. If they are the same, I remove (i+). However the problem is that my output are printing according to my loop..Just need to get out of the loop .