Update Largest.java
The idea is to check whether each element in the array is of length 1. If that's not the case, add all 1 length element to one array while trying to split the greater than 1 elements. So far I'm still struggling to split the > 1 elements, my loop can only split the second from last element