BucketList
Write a program called 'BucketList.java' that accepts as input (i) the name of a text file containing categories, one per line, and (ii) a comma-separated list of words to categorise. The program will output a list of the words in each category followed by a list of those words that could not be assigned to any.